requirements.txt
numpy
scipy
scikit-image
matplotlib
pillow
tqdm
tensorflow
keras
flask